requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu113
torch
torchvision
safetensors
huggingface_hub
pillow
gradio
spaces
tensorboard